Move Byte BPE loading from `GPTNeoXTokenizer` to `ByteBPETokenizer`
